+++ libyul/backends/evm/AbstractAssembly.h
@@ -110,6 +110,9 @@ public:
 	/// Appends an assignment to an immutable variable.
 	virtual void appendImmutableAssignment(std::string const& _identifier) = 0;
 
+	/// Appends data to the very end of the bytecode. Repeated calls concatenate.
+	virtual void appendToAuxiliaryData(bytes const& _data) = 0;
+
 	/// Mark this assembly as invalid. Any attempt to request bytecode from it should throw.
 	virtual void markAsInvalid() = 0;
 };